Updated Shader Code

+++ blue-beige.frag
@@ -1,4 +1,7 @@
-precision highp float;
+/* Blue-Beige */
+/* https://glslsandbox.com/e#67474.0 */
+
+precision mediump float;
 
 uniform vec2 resolution;
 uniform float time;
@@ -74,6 +77,3 @@ void main() {
 
     gl_FragColor = vec4(color, 1.);
 }
-
-/* Blue-Beige */
-/* https://glslsandbox.com/e#67474.0 */
